+++ sql/event_crypt.h
@@ -6,8 +6,8 @@
 #include "rpl_constants.h"
 #include "binlog_event.h"
 
-bool encrypt_event(uint32 offs, const Binlog_crypt_data *crypto, uchar* buf, uchar *ebuf, uint buf_len);
-bool decrypt_event(uint32 offs, const Binlog_crypt_data *crypto, uchar* buf, uchar *ebuf, uint buf_len);
+bool encrypt_event(uint32 offs, const Binlog_crypt_data *crypto, uchar* buf, uchar *ebuf, size_t buf_len);
+bool decrypt_event(uint32 offs, const Binlog_crypt_data *crypto, uchar* buf, uchar *ebuf, size_t buf_len);
 
 class Event_encrypter
 {
